+++ contracts/mock/uniswapv2/interfaces/IUniswapV2ERC20.sol
@@ -3,8 +3,8 @@
 pragma solidity >=0.5.0;
 
 interface IUniswapV2ERC20 {
-    event Approval(address indexed owner, address indexed spender, uint256 value);
-    event Transfer(address indexed from, address indexed to, uint256 value);
+    event Approval(address indexed owner, address indexed spender, uint value);
+    event Transfer(address indexed from, address indexed to, uint value);
 
     function name() external pure returns (string memory);
 
@@ -12,33 +12,33 @@ interface IUniswapV2ERC20 {
 
     function decimals() external pure returns (uint8);
 
-    function totalSupply() external view returns (uint256);
+    function totalSupply() external view returns (uint);
 
-    function balanceOf(address owner) external view returns (uint256);
+    function balanceOf(address owner) external view returns (uint);
 
-    function allowance(address owner, address spender) external view returns (uint256);
+    function allowance(address owner, address spender) external view returns (uint);
 
-    function approve(address spender, uint256 value) external returns (bool);
+    function approve(address spender, uint value) external returns (bool);
 
-    function transfer(address to, uint256 value) external returns (bool);
+    function transfer(address to, uint value) external returns (bool);
 
     function transferFrom(
         address from,
         address to,
-        uint256 value
+        uint value
     ) external returns (bool);
 
     function DOMAIN_SEPARATOR() external view returns (bytes32);
 
     function PERMIT_TYPEHASH() external pure returns (bytes32);
 
-    function nonces(address owner) external view returns (uint256);
+    function nonces(address owner) external view returns (uint);
 
     function permit(
         address owner,
         address spender,
-        uint256 value,
-        uint256 deadline,
+        uint value,
+        uint deadline,
         uint8 v,
         bytes32 r,
         bytes32 s
