+++ contracts/mock/uniswapv2/interfaces/IUniswapV2Factory.sol
@@ -3,7 +3,7 @@
 pragma solidity >=0.5.0;
 
 interface IUniswapV2Factory {
-    event PairCreated(address indexed token0, address indexed token1, address pair, uint256);
+    event PairCreated(address indexed token0, address indexed token1, address pair, uint);
 
     function feeTo() external view returns (address);
 
@@ -13,9 +13,9 @@ interface IUniswapV2Factory {
 
     function getPair(address tokenA, address tokenB) external view returns (address pair);
 
-    function allPairs(uint256) external view returns (address pair);
+    function allPairs(uint) external view returns (address pair);
 
-    function allPairsLength() external view returns (uint256);
+    function allPairsLength() external view returns (uint);
 
     function createPair(address tokenA, address tokenB) external returns (address pair);
