+++ contracts/mock/uniswapv2/libraries/TransferHelper.sol
@@ -7,9 +7,9 @@ library TransferHelper {
     function safeApprove(
         address token,
         address to,
-        uint256 value
+        uint value
     ) internal {
-        // bytes4(keccak256(bytes('approve(address,uint256)')));
+        // bytes4(keccak256(bytes('approve(address,uint)')));
         (bool success, bytes memory data) = token.call(
             abi.encodeWithSelector(0x095ea7b3, to, value)
         );
@@ -22,9 +22,9 @@ library TransferHelper {
     function safeTransfer(
         address token,
         address to,
-        uint256 value
+        uint value
     ) internal {
-        // bytes4(keccak256(bytes('transfer(address,uint256)')));
+        // bytes4(keccak256(bytes('transfer(address,uint)')));
         (bool success, bytes memory data) = token.call(
             abi.encodeWithSelector(0xa9059cbb, to, value)
         );
@@ -38,9 +38,9 @@ library TransferHelper {
         address token,
         address from,
         address to,
-        uint256 value
+        uint value
     ) internal {
-        // bytes4(keccak256(bytes('transferFrom(address,address,uint256)')));
+        // bytes4(keccak256(bytes('transferFrom(address,address,uint)')));
         (bool success, bytes memory data) = token.call(
             abi.encodeWithSelector(0x23b872dd, from, to, value)
         );
@@ -50,7 +50,7 @@ library TransferHelper {
         );
     }
 
-    function safeTransferETH(address to, uint256 value) internal {
+    function safeTransferETH(address to, uint value) internal {
         (bool success, ) = to.call{value: value}(new bytes(0));
         require(success, "TransferHelper: ETH_TRANSFER_FAILED");
     }
